requirements.txt
langchain
streamlit
faiss-cpu
python-dotenv
PyPDF2
langchain-google-genai